backend/requirements.txt
# FastAPI & server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# LangChain & LangGraph (all LLM calls go through LangChain)
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# HTTP client and search for tools
httpx>=0.27.0
ddgs>=9.0.0

# Config & validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0

# Code quality (PEP 8, formatting)
ruff>=0.8.0
